Optimize DMA according to GBA spec

+++ include/graphics.h
@@ -4910,7 +4910,7 @@ extern const u16 gTradeMenu_Pal[];
 extern const u8 gTradeMenu_Gfx[];
 extern const u16 gTradeMenu_Tilemap[];
 extern const u16 gTradeGba2_Pal[];
-extern const u8 gTradeGba_Gfx[];
+extern const u16 gTradeGba_Gfx[];
 extern const u16 gTradeMenuMonBox_Tilemap[];
 extern const u8 gTradeCursor_Gfx[];
 extern const u16 gTradeCursor_Pal[];
